backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "django>=6.0.4",
    "djangorestframework>=3.17.1",
    "httpx>=0.28.1",
    "psycopg2-binary>=2.9.12",
    "python-decouple>=3.8",
]
